requirements.txt
# Core IA
google-generativeai>=0.8.0
langchain>=0.3.0
langchain-google-genai>=2.0.0

# Data Science (versões com wheels pré-compilados)
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.5.0
faker>=33.0.0

# Visualização
streamlit>=1.40.0
plotly>=5.24.0
matplotlib>=3.9.0
seaborn>=0.13.0

# MLOps
mlflow>=2.18.0
python-dotenv>=1.0.0

# Testes
pytest>=8.3.0
pytest-cov>=6.0.0

# Utils
requests>=2.32.0
pydantic>=2.10.0
joblib>=1.4.0
